add entry page with input name and color and redirection to game
