+++ client/entry.html
@@ -8,21 +8,26 @@
 		<title>Warcaby</title>
 
 		<link rel="stylesheet" href="{{ STATIC_URL }}styles.css" />
-
-
 		<script src="{{ STATIC_URL }}angular.min.js"></script>
-		<script src="{{ STATIC_URL }}bower_components/jquery/dist/jquery.min.js"></script>
+		<script type="text/javascript" src="{{ STATIC_URL }}checkers_modules.js" data-cover></script>
 
-		<!-- <script src="{{ STATIC_URL }}angular.min.js"></script> -->
-
-		<!-- <link rel="stylesheet" type="text/css" href="bower_components/chessboard-js/css/chessboard.min.css" /> -->
-
+		<link rel="icon" href="icon.ico" type="image/x-icon" />
+		<link rel="shortcut icon" href="icon.ico" type="image/x-icon" />
 
 	</head>
     <body>
-		<center><br><br><br>
-			<button type="button" href="/play" id="button_id">GRAJ</button><br>
-		</center>
+		<div id="entry" class="container" ng-controller="codeController">
+			<center><br>
+				<input type="text" id="userNameText" placeholder="Nazwa gracza" maxlength="10">
+				<br>
+				<input type="radio" id="userColor1" name="userColors" value="white", checked>
+				<label for="userColor1">Białe</label>
+				<input type="radio" id="userColor2" name="userColors" value="black">
+				<label for="userColor1">Czarne</label>
+				<br>
+				<button type="button" onclick="angular.element(this).scope().loadGame()" id="button_id">GRAJ</button><br>
+			</center>
+		</div>
 	<br>
     </body>
 </html>
