+++ client/styles.css
@@ -46,55 +46,77 @@ div.coded_error{
 	background-color:#ffa1a1;
 }
 
-.clearfix-7da63 {
+.column {
+	float: left;
+	text-align: center;
+	vertical-align: middle;
+}
+
+.left, .right {
+	width: 25%;
+}
+
+.middle {
+	width: 50%;
+	margin: 0 auto;
+}
+
+/* Clear floats after the columns */
+.row:after {
+	content: "";
+	display: table;
 	clear: both;
   }
-  
-  .board-b72b1 {
-	border: 2px solid #404040;
-	box-sizing: content-box;
-  }
-  
-  .square-55d63 {
-	float: left;
-	position: relative;
-  
-	/* disable any native browser highlighting */
-	-webkit-touch-callout: none;
-	  -webkit-user-select: none;
-	   -khtml-user-select: none;
-		 -moz-user-select: none;
-		  -ms-user-select: none;
-			  user-select: none;
-  }
-  
-  .white-1e1d7 {
-	background-color: #f0d9b5;
-	color: #b58863;
-  }
-  
-  .black-3c85d {
-	background-color: #b58863;
-	color: #f0d9b5;
-  }
-  
-  .highlight1-32417, .highlight2-9c5d2 {
-	box-shadow: inset 0 0 3px 3px yellow;
-  }
-  
-  .notation-322f9 {
-	cursor: default;
-	font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-	font-size: 14px;
-	position: absolute;
-  }
-  
-  .alpha-d2270 {
-	bottom: 1px;
-	right: 3px;
-  }
-  
-  .numeric-fc462 {
-	top: 2px;
-	left: 2px;
-  }
+
+.clearfix-7da63 {
+	clear: both;
+}
+
+.board-b72b1 {
+border: 2px solid #404040;
+box-sizing: content-box;
+}
+
+.square-55d63 {
+float: left;
+position: relative;
+
+/* disable any native browser highlighting */
+-webkit-touch-callout: none;
+	-webkit-user-select: none;
+	-khtml-user-select: none;
+		-moz-user-select: none;
+		-ms-user-select: none;
+			user-select: none;
+}
+
+.white-1e1d7 {
+background-color: #f0d9b5;
+color: #b58863;
+}
+
+.black-3c85d {
+background-color: #b58863;
+color: #f0d9b5;
+}
+
+.highlight1-32417, .highlight2-9c5d2 {
+box-shadow: inset 0 0 3px 3px yellow;
+}
+
+.notation-322f9 {
+cursor: default;
+font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
+font-size: 14px;
+position: absolute;
+}
+
+.alpha-d2270 {
+bottom: 1px;
+right: 3px;
+}
+
+.numeric-fc462 {
+top: 2px;
+left: 2px;
+}
